refactored App.js to have most things rendered in PostsPage.js
